Final Touches
So, this should be basically the entire game completed. Lily did a portion of the betting system and buttons which is marked in the code, I'm just submitting it since I was debugging it.

I also did basic debugging, if you guys can find any problems just let me know, but otherwise I think we're done.